Need to check if repo exists first

+++ backups/backup.sh
@@ -22,20 +22,22 @@ for repo in "${repositories[@]}"; do
 
   echo "Cloning repository ${repository}"
   clone_url="https://${GH_TOKEN}@github.com/${organization}/${repository}.git"
-  git clone --mirror "${clone_url}"
+  git clone --mirror "${clone_url}" || true # Ignore errors here if repo does not exist
 
-  echo "Copying ${repository} to S3"
-  tar czf "${repository}.git.tar.gz" "${repository}.git"
-  aws s3 mv "${repository}.git.tar.gz" "${AWS_S3_BUCKET}/${repository}.git.tar.gz"
-  rm -r "${repository}.git"
+  if [ -d "${repository}.git" ]; then
+    echo "Copying ${repository} to S3"
+    tar czf "${repository}.git.tar.gz" "${repository}.git"
+    aws s3 mv "${repository}.git.tar.gz" "${AWS_S3_BUCKET}/${repository}.git.tar.gz"
+    rm -r "${repository}.git"
 
-  echo "Checking for existence of a wiki repo for ${repository}"
-  clone_url="https://${GH_TOKEN}@github.com/${organization}/${repository}.wiki.git"
-  git clone --mirror "${clone_url}" || true # Ignore errors here if wiki does not exist
-  if [ -d "${repository}.wiki.git" ]; then
-	echo "Copying ${repository}.wiki to S3"
-	tar czf "${repository}.wiki.git.tar.gz" "${repository}.wiki.git"
-	aws s3 mv "${repository}.wiki.git.tar.gz" "${AWS_S3_BUCKET}/${repository}.wiki.git.tar.gz"
-	rm -r "${repository}.wiki.git"
+    echo "Checking for existence of a wiki repo for ${repository}"
+    clone_url="https://${GH_TOKEN}@github.com/${organization}/${repository}.wiki.git"
+    git clone --mirror "${clone_url}" || true # Ignore errors here if wiki does not exist
+    if [ -d "${repository}.wiki.git" ]; then
+      echo "Copying ${repository}.wiki to S3"
+      tar czf "${repository}.wiki.git.tar.gz" "${repository}.wiki.git"
+      aws s3 mv "${repository}.wiki.git.tar.gz" "${AWS_S3_BUCKET}/${repository}.wiki.git.tar.gz"
+      rm -r "${repository}.wiki.git"
+    fi
   fi
 done
